+++ lib/busybox.sh
@@ -2,18 +2,27 @@
 
 set -e
 
-BUSYBOXTMP=${DISTBUILDDIR}/tmp/busybox-${BUSYBOXVERSION}
+if [ "${DISTBUILDVERBOSE}" == "1" ]; then
+  set -x
+fi
 
-rm -rf ${BUSYBOXTMP}
+if [ "${FORCEREBUILD}" == "1"  ]; then
+  rm -rf ${BUSYBOXTMP}
+fi
 
 download_and_verify ${BUSYBOXURL}/busybox-${BUSYBOXVERSION}.tar.bz2 ${BUSYBOXTMP}.tar.bz2
-( cd ${DISTBUILDDIR}/tmp && tar xjf ${BUSYBOXTMP}.tar.bz2 )
+
+if [ ! -d "${BUSYBOXTMP}" ]; then
+  ( cd ${DISTBUILDDIR}/tmp && tar xjf ${BUSYBOXTMP}.tar.bz2 )
+fi
+
 ( cd ${BUSYBOXTMP} && make CROSS_COMPILE="${DISTTARGET}-" defconfig && make ${GCCPARALLEL} CROSS_COMPILE="${DISTTARGET}-" && make CROSS_COMPILE="${DISTTARGET}-" CONFIG_PREFIX="${FAKEROOTDIR}" install )
 
 cp -v ${BUSYBOXTMP}/examples/depmod.pl ${FAKEROOTDIR}/cross-tools/bin
 chmod 755 ${FAKEROOTDIR}/cross-tools/bin/depmod.pl
 
-cat > ${FAKEROOTDIR}/etc/init.d/rcS << "EOF"
+mkdir -p ${FAKEROOTDIR}/etc/rc.d
+cat > ${FAKEROOTDIR}/etc/rc.d/startup << "EOF"
 #!/bin/sh
 
 # Start all init scripts in /etc/init.d
@@ -28,5 +37,5 @@ for i in /etc/init.d/S??* ;do
 done
 EOF
 
-chown 0:0 ${FAKEROOTDIR}/etc/init.d/rcS
-chmod u+x ${FAKEROOTDIR}/etc/init.d/rcS
+chown 0:0 ${FAKEROOTDIR}/etc/rc.d/startup
+chmod u+x ${FAKEROOTDIR}/etc/rc.d/startup
